+++ src/app/app.component.ts
@@ -7,7 +7,8 @@ import { FormControl } from '@angular/forms';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
-
+  public phone: number = 123456789;
+  public customMaska: [string, pattern];
 
   public form: FormControl;
   public form1: FormControl;
@@ -15,6 +16,13 @@ export class AppComponent {
   public cpfFormControl: FormControl;
   public clearIfNotMatch: FormControl;
   public numberOrStringForm: FormControl;
+  public sufixForm: FormControl;
+  public repeatForm: FormControl;
+
+  public pattern: pattern =  {
+    'P': {
+        pattern: new RegExp('\\d'),
+    }};
 
   public numberOrStringFormModel: string | number = '';
   public clearIfNotMatchModel: string | number = '';
@@ -22,16 +30,28 @@ export class AppComponent {
   public cpfModel: string | number = '';
   public dateModel: string | number = '';
   public showMaskModel: string | number = '';
+  public suf: string = '+7';
+  public sufixModel: string | number;
+  public repeatFormModel: string | number;
 
 
-  public constructor(
-  ) {
+  public constructor() {
     this.form = new FormControl('');
     this.form1 = new FormControl('');
     this.formDate = new FormControl('');
     this.cpfFormControl = new FormControl();
     this.clearIfNotMatch = new FormControl();
     this.numberOrStringForm = new FormControl();
+    this.sufixForm = new FormControl('');
+    this.repeatForm = new FormControl('');
+
+    this.customMaska = ['PPP-PPP-PPP', this.pattern];
   }
 
 }
+
+type pattern = {
+    [character: string]: {
+        pattern: RegExp
+    }
+};