+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+<a name="6.5.15"></a>
+# [6.5.14 support for repeat mask] (2018-10-08)
+You can pass into mask pattern with brackets
+
+<a name="6.4.14"></a>
+# [6.4.14 fixed '*' pattern in maskExpression ] (2018-10-05)
+Fixed pattern '*'
+
+<a name="6.4.13"></a>
+# [6.4.13 selection of input value and deletion of it is fixed ] (2018-10-05)
+You can select and delete value.
+
+<a name="6.4.12"></a>
+# [6.4.12 pipe with mask expression and custom Pattern ] (2018-04-10)
+You can pass array of expression and custom Pattern to pipe
+
+<a name="6.3.11"></a>
+# [6.3.11 fixed mask visable ] (2018-03-10)
+Fixed when page is loaded the mask is visible
+
+<a name="6.3.10"></a>
+# [6.3.10 fixed writeValue] (2018-03-10)
+Fixed writeValue in directive
+
+<a name="6.2.8"></a>
+# [6.2.8 support for Mask Pipe/MaskService] (2018-01-10)
+Added ability to use MaskPipe/MaskService inside component
+
+<a name="6.2.5"></a>
+# [6.2.5 support for IE] (2018-01-10)
+Added poliffyls for IE
+
 <a name="6.2.4"></a>
 # [6.2.4 support for mask] (2018-10-09)